Add https support to the OpenBSD installer.
The sets are downloaded using https per default. If the server does
not support https, the installer offers falling back to http after
asking for confirmation.

NOTE: Depending on the setup, people using autoinstall(8) might
need to add this new installer question to the response file.

   Unable to connect using https. Use http instead = yes|no

Joint work with and 'go for it' from deraadt@
Tested from and OK tb@
Feedback and OK for the approach halex@