Move vioscsireg.h to sys/dev/pv/vioscsireg.h
virtio itself is not PCI, it is a paravirtual bus on top of either PCI
or MMIO (arm).  This is the second step of moving the virtio files to
dev/pv.  No functional change.

Moving the file is done by removing and adding it again - the old
history can be found in the Attic of sys/dev/pci/vioscsireg.h.

no objections from sf@ mlarkin@