add m_pool_init(), a wrapper around pool_init for mbuf clusters.
m_pool_init is basically a call to pool_init with everythign except
the size and alignment specified, and a call to pool_set_constraints
so the memroy is always dma reachable. it also wires up the memory
with the custom mbuf pool allocator.

ok bluhm@ as part of a larger diff