Change priq enqueue policy to drop lower priority packets
The new priority queueing enqueue policy is such that when the
aggregate queue depth of an outgoing queue is exceeded we attempt
to find a non-empty queue of packets with lower priority than the
priority of a packet we're trying to enqueue and if there's such
queue, we drop the first packet from it.

This ensures that high priority traffic will almost always find
the place on the queue and low priority bulk traffic gets a better
chance at regulating its throughput.  There's no change in the
behavior if altered priorities are not used (e.g. via "set prio"
Pf directive, VLAN priorities and so on).

With a correction from dlg@, additional tests by dhill@
OK bluhm, mpi