The default for the "Server directory?" question can possibly come from either what information is extracted from the cgi server or from installurl(5). Otherwise a sane default is used.
Based on what server (HTTP_SERVER) is provided by the user decide
on what source to choose from for the default.

At the end of install_http() use the url from the cgi server as the
base for what's written to /etc/installurl if an official mirror was
used. Otherwise trim _url_base and remove the architecture and
snapshots or version part.

This fixes the problem reported by phessler@ which exposed how
fragile the current logic for this was after recent changes.

At this time of the release cycle the kernel presents itself as
release kernel, but we're still pre-release and the sets are still
in the snapshots directory on the mirrors. This was confusing the
installer script.

Thanks to phessler@ for finding this and testing.
Special thanks to tb@ who imposed on himself to try to understand
and review the diffs.

OK tb@, phessler@ (on a similar diff)
'commit when your are happy' deraadt@