Use memory barriers to prevent pointer use before initialization.
This work was sparked by the topic posted on hn by wuch. I am still not
sure that this fixes the defect he claims to have observed because I was
not able to create a proper regress test for it to manifest.

To that end, a proof of concept is more than welcomed!
Thank you for the report!

Discussed with and OK kettenis@, tedu@.