Add compressed fonts support in the kernel.
The wsdisplay_font structure has been modified to add two new members
(zdata and zdata_len) to store compressed font data and its size. We
define compressed fonts by setting the data field to NULL and populating
the zdata and zdata_len fields.

In wsfont_lock(), we check if the selected font needs to be inflated,
and we call the newly introduced wsfont_inflate() if required.

OK kettenis@