add a guard page on the end of each threads kernel stack.
this protects the process control block. it also makes the kernel
fault hard when the guard page gets hit by the stack, allowing
debugging of the problem set of function calls at the time the
problem occurs. without this, a big stack can overwrite the pcb,
which makes the thread generate a fault when it is switched back
onto a cpu, which is way too late.

discussed with guenther@
ok deraadt@