Add bwfm(4), a driver for Broadcom FullMAC WiFi controllers.  The FullMAC, in comparison to SoftMAC, does most WiFi handling in the firmware that's running on the controller.  This means we have to work around the net80211 stack while still implementing all the WiFi interfaces to userland.
This driver is still in early development.  So far it can connect
to open WiFis over the USB bus.  SDIO and PCIe support, for devices
like the Raspberry Pi 3 or the Macbooks, is not yet implemented.
Also mbufs on the transmit path leak and are not yet freed.

ok stsp@