rework transmit to get rid of ifq_deq_begin, and to improve speed.
i had a diff that did a small change to replace
ifq_deq_begin/commit/rollback with ifq_dequeue, and Eduard Nicodei
tweaked it a bit with Artturi Alm to use the fifo registers properly
and bump the ifq len.

the latter changes improve performance significantly.

"if you think that diff is correct, commit it" kettenis@