Fix 11g ifmedia modes in ath(4) (shown by 'ifconfig ath0 media').
There was code which set them up but didn't include 11b channels which
are part of 11g. And there was a hack which effectively stripped important
flag bits away and wouldn't work for modes with overlapping channels (b/g).
As a result, some flags were missing from 11g channels the driver reported
to net80211, which skipped over those channels when building the media list.

This gets us one step closer to supporting the AR5424.

ok mpi@